requirements.txt
django>=4.2.0
requests>=2.28.0
python-dotenv>=1.0.0
djangorestframework
psycopg2-binary>=2.9.3
gunicorn>=20.1.0